monitoring route

+++ lib/handle-drone.ts
@@ -3,13 +3,25 @@
 import { Drone, PrismaClient } from "@prisma/client";
 import { deleteObject, ref } from "firebase/storage";
 import { storage } from "./firebase.config";
+import { currentUser } from "@clerk/nextjs/server";
 
 const prisma = new PrismaClient();
 
 export async function GetDrones(): Promise<Drone[] | undefined> {
   try {
-    const farmers = await prisma.drone.findMany();
-    return farmers as any;
+    const drones = await prisma.drone.findMany();
+    return drones as any;
+  } catch (error) {
+    console.log(error);
+  } finally {
+    await prisma.$disconnect();
+  }
+}
+
+export async function GetDrone(id: number): Promise<Drone | undefined> {
+  try {
+    const drone = await prisma.drone.findFirst({ where: { id: id } });
+    return drone as any;
   } catch (error) {
     console.log(error);
   } finally {
@@ -19,30 +31,35 @@ export async function GetDrones(): Promise<Drone[] | undefined> {
 
 export async function deleteDrone(id: number): Promise<boolean> {
   try {
-    const drone = await prisma.drone.findUnique({ where: { id } });
-    if (drone) {
-      const assignedFarmer = await prisma.farmer.findFirst({
-        where: { droneId: drone.id },
-      });
-
-      if (assignedFarmer) {
-        await prisma.farmer.update({
-          where: { id: assignedFarmer.id },
-          data: {
-            droneId: null,
-          },
+    const session = await currentUser();
+    if (session) {
+      const drone = await prisma.drone.findUnique({ where: { id } });
+      if (drone) {
+        const assignedFarmer = await prisma.farmer.findFirst({
+          where: { droneId: drone.id },
         });
-      }
 
-      if (drone.image && drone.imgPath) {
-        const desertRef = ref(storage, drone.imgPath);
-        await deleteObject(desertRef);
+        if (assignedFarmer) {
+          await prisma.farmer.update({
+            where: { id: assignedFarmer.id },
+            data: {
+              droneId: null,
+            },
+          });
+        }
+
+        if (drone.image && drone.imgPath) {
+          const desertRef = ref(storage, drone.imgPath);
+          await deleteObject(desertRef);
+        }
+        await prisma.drone.delete({ where: { id } });
       }
-      await prisma.drone.delete({ where: { id } });
+      return true;
+    } else {
+      throw new Error("Unauthorized!");
     }
-    return true;
-  } catch (error) {
-    console.log(error);
+  } catch (error: any) {
+    throw new Error(error.message);
   } finally {
     await prisma.$disconnect();
     return false;
